refactor: update old `coins-menu adapter` -> `bookmark adapter`
