+++ app/src/main/java/com/hxl/cryptonumismatist/base/BaseAdapter.java
@@ -4,6 +4,7 @@ import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.AsyncListDiffer;
+import androidx.recyclerview.widget.DiffUtil;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.hxl.cryptonumismatist.util.EspressoIdlingResource;
@@ -13,10 +14,14 @@ import java.util.function.Function;
 
 public abstract class BaseAdapter<T, VH extends RecyclerView.ViewHolder & Function<T, Void>> extends RecyclerView.Adapter<VH> {
 
-    protected AsyncListDiffer<T> differ;
+    protected final AsyncListDiffer<T> differ;
 
     public List<T> getList() {
         return differ.getCurrentList();
+    }
+
+    public BaseAdapter(DiffUtil.ItemCallback<T> comparator) {
+        this.differ = new AsyncListDiffer<>(this, comparator);
     }
 
     public void setList(List<T> list) {
